pyproject.toml
[project]
name = "merlincli"
version = "0.1.0"
description = "BTC/USD market trend analysis CLI and dashboard"
authors = [ { name = "Merlin CLI" } ]
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.1",
    "numpy>=1.24",
    "requests>=2.31",
    "ccxt>=4.1",
    "pandas-ta>=0.3.14b0",
    "plotly>=5.18",
    "streamlit>=1.32",
    "click>=8.1",
    "python-dotenv>=1.0",
    "vaderSentiment>=3.3.2",
    "openai>=1.47",
    "google-generativeai>=0.5.0",
    "tenacity>=8.2"
]

[project.optional-dependencies]
dev = ["pytest", "black", "ruff"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
